Np 47593 handle fetching unpublished results (#6251)
* Create new hook for fetching registration

* Use new hook everywhere we fetch an registration

* Allow fetching unpublished result for duplicate check

* Add descriptive comment
