+++ src/pages/basic_data/app_admin/central_import/CentralImportCandidateMerge.tsx
@@ -6,12 +6,8 @@ import { getLanguageByUri } from 'nva-language';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { Redirect, useHistory, useParams } from 'react-router-dom';
-import {
-  fetchImportCandidate,
-  fetchRegistration,
-  updateImportCandidateStatus,
-  updateRegistration,
-} from '../../../../api/registrationApi';
+import { useFetchRegistration } from '../../../../api/hooks/useFetchRegistration';
+import { fetchImportCandidate, updateImportCandidateStatus, updateRegistration } from '../../../../api/registrationApi';
 import { PageSpinner } from '../../../../components/PageSpinner';
 import { setNotification } from '../../../../redux/notificationSlice';
 import { AssociatedLink } from '../../../../types/associatedArtifact.types';
@@ -42,11 +38,7 @@ export const CentralImportCandidateMerge = () => {
   const history = useHistory<BasicDataLocationState>();
   const { candidateIdentifier, registrationIdentifier } = useParams<MergeImportCandidateParams>();
 
-  const registrationQuery = useQuery({
-    queryKey: ['registration', registrationIdentifier],
-    queryFn: () => fetchRegistration(registrationIdentifier),
-    meta: { errorMessage: t('feedback.error.get_registration') },
-  });
+  const registrationQuery = useFetchRegistration(registrationIdentifier);
 
   const importCandidateQuery = useQuery({
     queryKey: ['importCandidate', candidateIdentifier],
