+++ src/pages/basic_data/app_admin/central_import/CentralImportDuplicationCheckPage.tsx
@@ -4,7 +4,8 @@ import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { Link, useLocation, useParams } from 'react-router-dom';
-import { fetchImportCandidate, fetchRegistration, updateImportCandidateStatus } from '../../../../api/registrationApi';
+import { useFetchRegistration } from '../../../../api/hooks/useFetchRegistration';
+import { fetchImportCandidate, updateImportCandidateStatus } from '../../../../api/registrationApi';
 import { FetchImportCandidatesParams, fetchImportCandidates } from '../../../../api/searchApi';
 import { ConfirmMessageDialog } from '../../../../components/ConfirmMessageDialog';
 import { PageSpinner } from '../../../../components/PageSpinner';
@@ -70,12 +71,9 @@ export const CentralImportDuplicationCheckPage = () => {
       ),
   });
 
-  const importedRegistrationId = importCandidate?.importStatus.nvaPublicationId ?? '';
-  const importedRegistrationQuery = useQuery({
-    queryKey: ['registration', importedRegistrationId],
-    enabled: importCandidate?.importStatus.candidateStatus === 'IMPORTED' && !!importedRegistrationId,
-    queryFn: () => fetchRegistration(getIdentifierFromId(importedRegistrationId)),
-    meta: { errorMessage: t('feedback.error.get_registration') },
+  const importedRegistrationIdentifier = getIdentifierFromId(importCandidate?.importStatus.nvaPublicationId ?? '');
+  const importedRegistrationQuery = useFetchRegistration(importedRegistrationIdentifier, {
+    enabled: importCandidate?.importStatus.candidateStatus === 'IMPORTED',
   });
 
   useEffect(() => {
