+++ src/pages/public_registration/action_accordions/DeletedRegistrationInformation.tsx
@@ -1,8 +1,7 @@
 import { Box, Link, Skeleton, Typography } from '@mui/material';
-import { useQuery } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
+import { useFetchRegistration } from '../../../api/hooks/useFetchRegistration';
 import { useFetchUserQuery } from '../../../api/hooks/useFetchUserQuery';
-import { fetchRegistration } from '../../../api/registrationApi';
 import { ProfilePicture } from '../../../components/ProfilePicture';
 import { PublicationNote, Registration } from '../../../types/registration.types';
 import { toDateString } from '../../../utils/date-helpers';
@@ -22,13 +21,7 @@ export const DeletedRegistrationInformation = ({
   const { t } = useTranslation();
 
   const duplicateRegistrationIdentifier = getIdentifierFromId(registration.duplicateOf ?? '');
-
-  const duplicateRegistrationQuery = useQuery({
-    enabled: !!duplicateRegistrationIdentifier,
-    queryKey: ['registration', duplicateRegistrationIdentifier],
-    queryFn: () => fetchRegistration(duplicateRegistrationIdentifier),
-    meta: { errorMessage: t('feedback.error.get_registration') },
-  });
+  const duplicateRegistrationQuery = useFetchRegistration(duplicateRegistrationIdentifier);
 
   const senderQuery = useFetchUserQuery(unpublishingNote.createdBy ?? '');
