+++ src/pages/registration/resource_type_tab/sub_type_forms/artistic_types/OutputRow.tsx
@@ -3,10 +3,9 @@ import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
 import CancelIcon from '@mui/icons-material/Cancel';
 import EditIcon from '@mui/icons-material/Edit';
 import { Box, Button, Skeleton, TableCell, TableRow, Tooltip, Typography } from '@mui/material';
-import { useQuery } from '@tanstack/react-query';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { fetchRegistration } from '../../../../../api/registrationApi';
+import { useFetchRegistration } from '../../../../../api/hooks/useFetchRegistration';
 import { ConfirmDialog } from '../../../../../components/ConfirmDialog';
 import {
   ArtisticOutputItem,
@@ -85,12 +84,7 @@ export const OutputRow = ({
   const shouldFetchItem = item.type === 'ExhibitionCatalog';
   const exhibitionCatalogIdentifier = shouldFetchItem && item.id ? getIdentifierFromId(item.id) : '';
 
-  const exhibitionCatalogQuery = useQuery({
-    enabled: !!exhibitionCatalogIdentifier,
-    queryKey: ['registration', exhibitionCatalogIdentifier],
-    queryFn: () => fetchRegistration(exhibitionCatalogIdentifier),
-    meta: { errorMessage: t('feedback.error.get_registration') },
-  });
+  const exhibitionCatalogQuery = useFetchRegistration(exhibitionCatalogIdentifier);
 
   const title = shouldFetchItem ? exhibitionCatalogQuery.data?.entityDescription?.mainTitle : getOutputName(item);
